Fix message arg type

+++ formats/ctf/ctf.c
@@ -987,7 +987,7 @@ int create_stream_packet_index(struct ctf_trace *td,
 
 		if (packet_index.packet_size > (filestats.st_size - packet_index.offset) * CHAR_BIT) {
 			fprintf(stdout, "[error] Packet size (%zu bits) is larger than remaining file size (%zu bits).\n",
-				packet_index.content_size, (filestats.st_size - packet_index.offset) * CHAR_BIT);
+				packet_index.content_size, (size_t) (filestats.st_size - packet_index.offset) * CHAR_BIT);
 			return -EINVAL;
 		}
